chore: move evalEnergy to API
